CrimesAnalysis_GCP/FINAL-ARCHI1/requirements.txt
functions-framework==3.*
google-cloud-bigquery==2.34.0
google-cloud-storage==1.42.0